requirements.txt
Celery
pillow
sorl-thumbnail